+++ src/budy/models/order.py
@@ -230,6 +230,20 @@ class Order(bundle.Bundle):
         )
 
     @classmethod
+    @appier.link(name="Inventory Report", context=True)
+    def inventory_url(cls, view=None, context=None, absolute=False):
+        return appier.get_app().url_for(
+            "order_api.inventory", view=view, context=context, absolute=absolute
+        )
+
+    @classmethod
+    @appier.link(name="Inventory PDF", context=True)
+    def inventory_pdf_url(cls, view=None, context=None, absolute=False):
+        return appier.get_app().url_for(
+            "order_api.inventory_pdf", view=view, context=context, absolute=absolute
+        )
+
+    @classmethod
     @appier.operation(
         name="Import CTT",
         parameters=(
@@ -1552,6 +1566,12 @@ class Order(bundle.Bundle):
             seeplus_updates=seeplus_updates,
         )
         self.save()
+
+    @appier.link(name="Report")
+    def report_url(self, absolute=False):
+        return appier.get_app().url_for(
+            "order_api.report", id=self.id, absolute=absolute
+        )
 
     @appier.link(name="Export Lines CSV")
     def lines_csv_url(self, absolute=False):